backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
mysql-connector-python
pyjwt
werkzeug
python-dotenv
fastapi
uvicorn
bytez
pydantic
sqlalchemy
